requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
requests>=2.32.0,<3.0.0
openai>=1.40.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
